Reorganise modules, have scripts separate to types, remove a bunch of unsed stuff
